# docs/openapi/cliser-data-node.openapi.current.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CLISER DATA NODE API",
    "version": "1.0.0",
    "description": "HTTP API do CLISER DATA NODE.\n\nAPI Version: v1\nBase Path: /api/v1"
  },
  "paths": {
...
    "/api/v1/namespaces": {
      "get": {
        "tags": [
          "namespaces"
        ],
        "summary": "List Namespaces",
        "operationId": "list_namespaces_api_v1_namespaces_get",
        "parameters": [
          {
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(ACTIVE|DISABLED)$",
              "title": "Status"
            },
            "name": "status",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            },
            "name": "authorization",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NamespaceListResponse"
                },
                "examples": {
                  "NamespaceListResponse": {
                    "$ref": "#/components/examples/NamespaceListResponse"
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                },
                "examples": {
                  "ErrorValidation": {
                    "$ref": "#/components/examples/ErrorValidation"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "namespaces"
        ],
        "summary": "Create Namespace",
        "operationId": "create_namespace_api_v1_namespaces_post",
        "parameters": [
          {
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            },
            "name": "authorization",
            "in": "header"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NamespaceCreateRequest"
              },
              "examples": {
                "NamespaceCreate": {
                  "$ref": "#/components/examples/NamespaceCreateRequest"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NamespaceResponse"
                },
                "examples": {
                  "NamespaceResponse": {
                    "$ref": "#/components/examples/NamespaceResponse"
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                },
                "examples": {
                  "ErrorValidation": {
                    "$ref": "#/components/examples/ErrorValidation"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v1/namespaces/{namespace}": {
      "get": {
        "tags": [
          "namespaces"
        ],
        "summary": "Get Namespace",
        "operationId": "get_namespace_api_v1_namespaces__namespace__get",
        "parameters": [
          {
            "required": true,
            "schema": {
              "type": "string",
              "title": "Namespace"
            },
            "name": "namespace",
            "in": "path"
          },
          {
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            },
            "name": "authorization",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NamespaceResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                },
                "examples": {
                  "ErrorValidation": {
                    "$ref": "#/components/examples/ErrorValidation"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v1/namespaces/{namespace}/enable": {
      "post": {
        "tags": [
          "namespaces"
        ],
        "summary": "Enable Namespace",
        "operationId": "enable_namespace_api_v1_namespaces__namespace__enable_post",
        "parameters": [
          {
            "required": true,
            "schema": {
              "type": "string",
              "title": "Namespace"
            },
            "name": "namespace",
            "in": "path"
          },
          {
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            },
            "name": "authorization",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NamespaceResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                },
                "examples": {
                  "ErrorValidation": {
                    "$ref": "#/components/examples/ErrorValidation"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v1/namespaces/{namespace}/disable": {
      "post": {
        "tags": [
          "namespaces"
        ],
        "summary": "Disable Namespace",
        "operationId": "disable_namespace_api_v1_namespaces__namespace__disable_post",
        "parameters": [
          {
            "required": true,
            "schema": {
              "type": "string",
              "title": "Namespace"
            },
            "name": "namespace",
            "in": "path"
          },
          {
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            },
            "name": "authorization",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NamespaceResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                },
                "examples": {
                  "ErrorValidation": {
                    "$ref": "#/components/examples/ErrorValidation"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v1/namespaces/{namespace}/objects": {
      "get": {
        "tags": [
          "objects"
        ],
        "summary": "List Objects",
        "operationId": "list_objects_api_v1_namespaces__namespace__objects_get",
        "parameters": [
          {
            "required": true,
            "schema": {
              "type": "string",
              "title": "Namespace"
            },
            "name": "namespace",
            "in": "path"
          },
          {
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            },
            "name": "authorization",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectListResponse"
                },
                "examples": {
                  "ObjectListResponse": {
                    "$ref": "#/components/examples/ObjectListResponse"
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                },
                "examples": {
                  "ErrorValidation": {
                    "$ref": "#/components/examples/ErrorValidation"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "objects"
        ],
        "summary": "Create Object",
        "operationId": "create_object_api_v1_namespaces__namespace__objects_post",
        "parameters": [
          {
            "required": true,
            "schema": {
              "type": "string",
              "title": "Namespace"
            },
            "name": "namespace",
            "in": "path"
          },
          {
            "required": false,
            "schema": {
              "type": "string",
              "title": "Idempotency-Key"
            },
            "name": "Idempotency-Key",
            "in": "header"
          },
          {
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            },
            "name": "authorization",
            "in": "header"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ObjectCreateRequest"
              },
              "examples": {
                "ObjectCreate": {
                  "$ref": "#/components/examples/ObjectCreateRequest"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectCreateResponse"
                },
                "examples": {
                  "ObjectCreateResponse": {
                    "$ref": "#/components/examples/ObjectCreateResponse"
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                },
                "examples": {
                  "ErrorValidation": {
                    "$ref": "#/components/examples/ErrorValidation"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/v1/namespaces/{namespace}/objects/{object_id}/metadata": {
      "get": {
        "tags": [
          "metadata"
        ],
        "summary": "Get Object Metadata",
        "operationId": "get_object_metadata_api_v1_namespaces__namespace__objects__object_id__metadata_get",
        "parameters": [
          {
            "required": true,
            "schema": {
              "type": "string",
              "title": "Namespace"
            },
            "name": "namespace",
            "in": "path"
          },
          {
            "required": true,
            "schema": {
              "type": "string",
              "title": "Object Id"
            },
            "name": "object_id",
            "in": "path"
          },
          {
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            },
            "name": "authorization",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectMetadataResponse"
                },
                "examples": {
                  "ObjectMetadataResponse": {
                    "$ref": "#/components/examples/ObjectMetadataResponse"
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                },
                "examples": {
                  "ErrorValidation": {
                    "$ref": "#/components/examples/ErrorValidation"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "NamespaceCreateRequest": {
        "properties": {
          "namespace": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Namespace"
          },
          "quota_bytes": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Quota Bytes",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "namespace"
        ],
        "title": "NamespaceCreateRequest"
      },
      "NamespaceListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/NamespaceResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "title": "NamespaceListResponse"
      },
      "NamespaceResponse": {
        "properties": {
          "namespace": {
            "type": "string",
            "title": "Namespace"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "quota_bytes": {
            "type": "integer",
            "title": "Quota Bytes"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "namespace",
          "status",
          "quota_bytes",
          "created_at",
          "updated_at"
        ],
        "title": "NamespaceResponse"
      },
      "ObjectCreateRequest": {
        "properties": {
          "data": {
            "type": "string",
            "minLength": 1,
            "title": "Data"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "ObjectCreateRequest"
      },
      "ObjectCreateResponse": {
        "properties": {
          "object_id": {
            "type": "string",
            "title": "Object Id"
          },
          "namespace": {
            "type": "string",
            "title": "Namespace"
          },
          "size": {
            "type": "integer",
            "title": "Size"
          },
          "content_hash": {
            "type": "string",
            "title": "Content Hash"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "object_id",
          "namespace",
          "size",
          "content_hash",
          "status"
        ],
        "title": "ObjectCreateResponse"
      },
...
      "ObjectListItem": {
        "properties": {
          "object_id": {
            "type": "string",
            "title": "Object Id"
          },
          "namespace": {
            "type": "string",
            "title": "Namespace"
          },
          "size": {
            "type": "integer",
            "title": "Size"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "object_id",
          "namespace",
          "size",
          "status"
        ],
        "title": "ObjectListItem"
      },
      "ObjectListResponse": {
        "properties": {
          "namespace": {
            "type": "string",
            "title": "Namespace"
          },
          "objects": {
            "items": {
              "$ref": "#/components/schemas/ObjectListItem"
            },
            "type": "array",
            "title": "Objects"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          }
        },
        "type": "object",
        "required": [
          "namespace",
          "objects",
          "count"
        ],
        "title": "ObjectListResponse"
      },
      "ObjectMetadataResponse": {
        "properties": {
          "object_id": {
            "type": "string",
            "title": "Object Id"
          },
          "namespace": {
            "type": "string",
            "title": "Namespace"
          },
          "size": {
            "type": "integer",
            "title": "Size"
          },
          "content_hash": {
            "type": "string",
            "title": "Content Hash"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "object_id",
          "namespace",
          "size",
          "content_hash",
          "status"
        ],
        "title": "ObjectMetadataResponse"
      },
...
      "APIError": {
        "title": "APIError",
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "code",
              "message",
              "request_id"
            ],
            "properties": {
              "code": {
                "type": "string",
                "example": "OBJECT_NOT_FOUND"
              },
              "message": {
                "type": "string",
                "example": "Objeto não encontrado"
              },
              "request_id": {
                "type": [
                  "string",
                  "null"
                ],
                "example": "req_01JEXAMPLE"
              },
              "details": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "credential",
        "description": "Credencial de API enviada no cabeçalho Authorization como Bearer <credential>."
      }
    },
    "examples": {
      "NamespaceCreateRequest": {
        "value": {
          "namespace": "documents",
          "quota_bytes": 10737418240
        },
        "summary": "Criação de namespace"
      },
      "NamespaceResponse": {
        "value": {
          "namespace": "documents",
          "status": "ACTIVE",
          "quota_bytes": 10737418240,
          "created_at": "2026-09-15T15:00:00Z",
          "updated_at": "2026-09-15T15:00:00Z"
        },
        "summary": "Namespace ativo"
      },
      "NamespaceListResponse": {
        "value": {
          "items": [
            {
              "namespace": "documents",
              "status": "ACTIVE",
              "quota_bytes": 10737418240,
              "created_at": "2026-09-15T15:00:00Z",
              "updated_at": "2026-09-15T15:00:00Z"
            }
          ],
          "total": 1
        },
        "summary": "Lista de namespaces"
      },
      "ObjectCreateRequest": {
        "value": {
          "data": "Q0xJU0VSIERBVEEgTk9ERQ=="
        },
        "summary": "Objeto em Base64",
        "description": "O campo data deve conter Base64 válido."
      },
      "ObjectCreateResponse": {
        "value": {
          "object_id": "obj_01JEXAMPLE",
          "namespace": "documents",
          "size": 16,
          "content_hash": "sha256:01JEXAMPLE",
          "status": "ACTIVE"
        },
        "summary": "Objeto criado"
      },
      "ObjectListResponse": {
        "value": {
          "namespace": "documents",
          "objects": [
            {
              "object_id": "obj_01JEXAMPLE",
              "namespace": "documents",
              "size": 16,
              "created_at": "2026-09-14T12:00:00Z",
              "status": "ACTIVE"
            }
          ],
          "count": 1
        },
        "summary": "Lista de objetos"
      },
      "ObjectMetadataResponse": {
        "value": {
          "object_id": "obj_01JEXAMPLE",
          "namespace": "documents",
          "size": 16,
          "content_hash": "sha256:01JEXAMPLE",
          "created_at": "2026-09-15T15:00:00Z",
          "status": "ACTIVE"
        },
        "summary": "Metadados do objeto"
      },
...
      "ErrorValidation": {
        "value": {
          "error": {
            "code": "VALIDATION_ERROR",
            "message": "Request validation failed.",
            "request_id": "req_01JEXAMPLE",
            "details": {
              "errors": [
                {
                  "type": "value_error",
                  "loc": [
                    "body",
                    "namespace"
                  ],
                  "msg": "field required"
                }
              ]
            }
          }
        },
        "summary": "Erro de validação"
      },
...
    }
  },
...
}